Add state to app and refactor files
