+++ src/Card/Card.js
@@ -1,3 +1,1 @@
 
-
-export default Card;
