+++ src/Favorites/Favorites.js
@@ -1,3 +1,1 @@
 
-
-export default Favorites;
