+++ src/GenX/GenX.js
@@ -1,3 +1,1 @@
 
-
-export default GenX;
