+++ src/SearchBar/SearchBar.js
@@ -1,3 +1,1 @@
 
-
-export default SearchBar;
